Remove Python query planner.
