+++ daft/plan_scheduler/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from daft.plan_scheduler.physical_plan_scheduler import PhysicalPlanScheduler
+
+__all__ = ["PhysicalPlanScheduler"]
