--- daft/planner/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from daft.planner.planner import PhysicalPlanScheduler
-
-__all__ = ["PhysicalPlanScheduler"]
